% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bank_hist.R
\name{bank_hist}
\alias{bank_hist}
\title{Get Historical Bank info from TBB}
\usage{
bank_hist()
}
\value{
A tibble
}
\description{
Get Historical Bank info from TBB
}
\examples{

\dontrun{
dt <- bank_hist()
}

}
