% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banks_financial.R
\name{banks_financial}
\alias{banks_financial}
\title{Get banks info for financial data from TBB}
\usage{
banks_financial()
}
\value{
A tibble
}
\description{
Get banks info for financial data from TBB
}
\examples{

\dontrun{
dt <- banks_financial()
}

}
