% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banks_info.R
\name{banks_info}
\alias{banks_info}
\title{Get list of banks info data from TBB}
\usage{
banks_info(date = format(Sys.time(), "\%d/\%m/\%Y"))
}
\arguments{
\item{date}{Optional Date in DD/MM/YY format}
}
\value{
A tibble
}
\description{
Get list of banks info data from TBB
}
\examples{

\dontrun{
dt <- banks_info()
}

}
