% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banks_summary.R
\name{banks_summary}
\alias{banks_summary}
\title{Get list of banks data from TBB}
\usage{
banks_summary()
}
\value{
A tibble
}
\description{
Get list of banks data from TBB
}
\examples{

\dontrun{
dt <- banks_summary()
}

}
