% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categories_risk.R
\name{categories_risk}
\alias{categories_risk}
\title{Get parameters and codes for risk data from TBB}
\usage{
categories_risk()
}
\value{
A tibble
}
\description{
Get parameters and codes for risk data from TBB
}
\examples{


\dontrun{
dt <- categories_risk()
}

}
