% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_financial.R
\name{data_financial}
\alias{data_financial}
\title{Get financial tables data from TBB}
\usage{
data_financial(periods, tables, banks = NULL, groups = NULL)
}
\arguments{
\item{periods}{Period codes from periods_financial() KEY column}

\item{tables}{Table codes from tables_financial() UNIQUE_KEY column}

\item{banks}{Bank codes from banks_financial() BANKA_KODU column}

\item{groups}{Group codes from groups_financial() GRUP_NO column}
}
\value{
A tibble
}
\description{
Get financial tables data from TBB
}
\examples{


\dontrun{
dt <- data_financial()
}

}
