% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_region.R
\name{data_region}
\alias{data_region}
\title{Get regional data from TBB}
\usage{
data_region(years = NULL, regions = NULL, parameters = NULL)
}
\arguments{
\item{years}{Years}

\item{regions}{Region Codes}

\item{parameters}{Data Codes}
}
\value{
A tibble
}
\description{
Get regional data from TBB
}
\examples{

\dontrun{
dt <- data_region(years = c(2000))
}

}
