% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_risk.R
\name{data_risk}
\alias{data_risk}
\title{Get parameters and codes for risk data from TBB}
\usage{
data_risk(categories = NULL, periods = NULL)
}
\arguments{
\item{categories}{Data Categories}

\item{periods}{Data Periods}
}
\value{
A tibble
}
\description{
Get parameters and codes for risk data from TBB
}
\examples{

\dontrun{
dt <- data_risk()
}

}
