% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_hist.R
\name{gm_hist}
\alias{gm_hist}
\title{Get Historical General Manager info from TBB}
\usage{
gm_hist()
}
\value{
A tibble
}
\description{
Get Historical General Manager info from TBB
}
\examples{

\dontrun{
dt <- gm_hist()
}

}
