% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups_financial.R
\name{groups_financial}
\alias{groups_financial}
\title{Get bank groups info for financial data from TBB}
\usage{
groups_financial()
}
\value{
A tibble
}
\description{
Get bank groups info for financial data from TBB
}
\examples{

\dontrun{
dt <- groups_financial()
}

}
