% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params_region.R
\name{params_region}
\alias{params_region}
\title{Get parameters and codes for regional data from TBB}
\usage{
params_region()
}
\value{
A tibble
}
\description{
Get parameters and codes for regional data from TBB
}
\examples{

\dontrun{
dt <- params_region()
}

}
