% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periods_financial.R
\name{periods_financial}
\alias{periods_financial}
\title{Get periods for financial data from TBB}
\usage{
periods_financial()
}
\value{
A tibble
}
\description{
Get periods for financial data from TBB
}
\examples{

\dontrun{
dt <- periods_financial()
}

}
