% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periods_risk.R
\name{periods_risk}
\alias{periods_risk}
\title{Get available years for risk data from TBB}
\usage{
periods_risk()
}
\value{
A tibble
}
\description{
Get available years for risk data from TBB
}
\examples{

\dontrun{
dt <- periods_risk()
}

}
