% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbat-package.R
\docType{package}
\name{tbat-package}
\alias{tbat}
\alias{tbat-package}
\title{tbat: Download Data From The Banks Association Of Turkey}
\description{
The Banks Association Of Turkey has a Data Query
    System but has no public API. This package uses POST queries to
    download data from the webpage.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/emraher/tbat}
  \item Report bugs at \url{https://github.com/emraher/tbat/issues}
}

}
\author{
\strong{Maintainer}: Emrah Er \email{eer@eremrah.com}

}
\keyword{internal}
