% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/years_region.R
\name{years_region}
\alias{years_region}
\title{Get available years for regional data from TBB
#'}
\usage{
years_region()
}
\value{
A tibble
}
\description{
Get available years for regional data from TBB
#'
}
\examples{

\dontrun{
dt <- years_region()
}

}
